FAC-219: Move LoadKitLastMajorModificationTimePort to core module
