+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/assessment/ReportAssessmentService.java
@@ -12,7 +12,6 @@ import org.flickit.assessment.core.application.domain.report.TopAttributeResolve
 import org.flickit.assessment.core.application.port.in.assessment.ReportAssessmentUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentReportInfoPort;
 import org.flickit.assessment.core.application.port.out.qualityattributevalue.LoadAttributeValueListPort;
-import org.flickit.assessment.kit.application.port.out.assessmentkit.LoadKitLastMajorModificationTimePort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
