+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/subject/ReportSubjectServiceTest.java
@@ -8,7 +8,6 @@ import org.flickit.assessment.core.application.domain.report.SubjectReport;
 import org.flickit.assessment.core.application.port.in.subject.ReportSubjectUseCase;
 import org.flickit.assessment.core.application.port.out.subject.LoadSubjectReportInfoPort;
 import org.flickit.assessment.core.test.fixture.application.MaturityLevelMother;
-import org.flickit.assessment.kit.application.port.out.assessmentkit.LoadKitLastMajorModificationTimePort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
